Allows top level errors to continue through
Previously top-level errors, which not only come from the server but
were also added as the fallback approach, would throw, be caught, and
call onError and return an empty markup for client side rendering. This
was ok but `afterResponse` wasn't running which would potentially impact
plugins that would use this to check for errors that happened in
`willRequest`, or `getViewData`.

Now when a top-level error is encountered we continue the flow,
`onError` is called for the error and then we proceed to iterate over
each job and call `afterResponse`.

If an error is encountered during that process then we terminate and
fallback to empty markup.